# general-api/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "General API",
    "description" : "Spring Boot API that delivers some general functionality to the users of AutoPacker."
  },
  "servers" : [ {
    "url" : "http://dev.libane.tk:8082"
  } ],
  "tags" : [ {
    "name" : "admins",
    "description" : "Operations available only to admins."
  }, {
    "name" : "members",
    "description" : "Operations available to all our members."
  } ],
  "paths" : {
    "/languages/new" : {
      "post" : {
        "tags" : [ "admins" ],
        "summary" : "Add a new supported language to the application",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "language", "version" ],
                "properties" : {
                  "language" : {
                    "type" : "string",
                    "example" : "java"
                  },
                  "version" : {
                    "type" : "integer",
                    "example" : 8
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Language successfully added"
          },
          "400" : {
            "description" : "Language creation failed"
          }
        }
      }
    },
    "/languages/all" : {
      "get" : {
        "tags" : [ "members" ],
        "description" : "Returns a list of all our supported languages and versions",
        "responses" : {
          "200" : {
            "description" : "Server retrieved",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Language"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Couldn't fetch list of languages"
          }
        }
      }
    }
  },
  "components" : {
    "securitySchemes" : {
      "authKey" : {
        "type" : "apiKey",
        "in" : "header",
        "name" : "Cookie"
      }
    },
    "responses" : {
      "403" : {
        "description" : "Access denied, token is not present or is invalid.",
        "content" : {
          "application/json" : {
            "schema" : {
              "type" : "object",
              "properties" : {
                "timestamp" : {
                  "type" : "string",
                  "example" : "2020-03-10T23:46:32.380+0000"
                },
                "status" : {
                  "type" : "integer",
                  "example" : 403
                },
                "error" : {
                  "type" : "string",
                  "example" : "Forbidden"
                },
                "message" : {
                  "type" : "string",
                  "example" : "Access Denied"
                },
                "path" : {
                  "type" : "string",
                  "example" : "/auth/login"
                }
              }
            }
          }
        }
      }
    },
    "schemas" : {
      "Language" : {
        "type" : "object",
        "description" : "Represents a language",
        "required" : [ "id", "language", "versions" ],
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 1
          },
          "language" : {
            "type" : "string",
            "example" : "Java"
          },
          "versions" : {
            "type" : "array",
            "example" : [ 8, 11 ],
            "items" : {
              "$ref" : "#/components/schemas/Version"
            }
          }
        }
      },
      "Version" : {
        "type" : "object",
        "description" : "Represents a language version",
        "required" : [ "id", "version", "language" ],
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 1
          },
          "version" : {
            "type" : "integer",
            "example" : 11
          },
          "language" : {
            "$ref" : "#/components/schemas/Language"
          }
        }
      }
    }
  }
}